third_party/blink/web_tests/fast/loader/target-blank-to-local-resource-expected.txt

CONSOLE ERROR: Not allowed to load local resource: chrome://downloads/
